+++ timeago.go
@@ -75,6 +75,8 @@ func calculateTheResult(seconds int) string {
 	switch {
 	case optionIsEnabled("online") && seconds < 60:
 		return trans().Online
+	case optionIsEnabled("justNow") && seconds < 60:
+		return trans().JustNow
 	case seconds < 60:
 		return getWords("seconds", seconds)
 	case minutes < 60:
@@ -115,8 +117,13 @@ func getWords(timeKind string, num int) string {
 	time := getTimeTranslations()
 
 	translation := time[timeKind][form]
+	result := strconv.Itoa(num) + " " + translation
 
-	return strconv.Itoa(num) + " " + translation + " " + trans().Ago
+	if optionIsEnabled("noSuffix") {
+		return result
+	}
+
+	return result + " " + trans().Ago
 }
 
 // Check if option was passed by a Parse function