+++ docs/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ----
 
+## v2.1.1 (2022-01-28)
+
+- Added option `justNow` that prints `Just now` if time is within 60 minutes;
+- Added option `noSuffix` that removes `ago` from the printed result;
+- Added more info to `OPTIONS.md` documentation file;
+
+----
+
 ## v2.1.0 (2022-01-27)
 
 - Changed: